ready text
